+++ environs.py
@@ -14,27 +14,31 @@ from pathlib import Path
 import marshmallow as ma
 from dotenv.main import load_dotenv, DotEnv, _walk_to_root
 
-__version__ = "5.1.0"
+__version__ = "6.0.0"
 __all__ = ["EnvError", "Env"]
 
 MARSHMALLOW_VERSION_INFO = tuple([int(part) for part in ma.__version__.split(".") if part.isdigit()])
 _PROXIED_PATTERN = re.compile(r"\s*{{\s*(\S*)\s*}}\s*")
 
-T = typing.TypeVar("T")
+_T = typing.TypeVar("_T")
+_StrType = str
+_BoolType = bool
+_IntType = int
+
 ErrorMapping = typing.Mapping[str, typing.List[str]]
 ErrorList = typing.List[str]
 FieldFactory = typing.Callable[..., ma.fields.Field]
-Subcast = typing.Union[typing.Type, typing.Callable[..., T]]
+Subcast = typing.Union[typing.Type, typing.Callable[..., _T]]
 FieldType = typing.Type[ma.fields.Field]
 FieldOrFactory = typing.Union[FieldType, FieldFactory]
-ParserMethod = typing.Callable[..., T]
+ParserMethod = typing.Callable[..., _T]
 
 
-class EnvError(Exception):
-    pass
+class EnvError(ValueError):
+    """Raised when an environment variable or if a required environment variable is unset."""
 
 
-class EnvValidationError(ValueError, EnvError):
+class EnvValidationError(EnvError):
     def __init__(self, message: str, error_messages: typing.Union[ErrorList, ErrorMapping]):
         self.error_messages = error_messages
         super().__init__(message)
@@ -44,12 +48,16 @@ class EnvSealedError(TypeError, EnvError):
     pass
 
 
+class ParserConflictError(ValueError):
+    """Raised when adding a custom parser that conflicts with a built-in parser method."""
+
+
 def _field2method(
     field_or_factory: FieldOrFactory, method_name: str, *, preprocess: typing.Callable = None
 ) -> ParserMethod:
     def method(
         self: "Env", name: str, default: typing.Any = ma.missing, subcast: Subcast = None, **kwargs
-    ) -> T:
+    ) -> _T:
         if self._sealed:
             raise EnvSealedError("Env has already been sealed. New values cannot be parsed.")
         missing = kwargs.pop("missing", None) or default
@@ -59,12 +67,11 @@ def _field2method(
             field = typing.cast(FieldFactory, field_or_factory)(subcast=subcast, missing=missing, **kwargs)
         parsed_key, raw_value, proxied_key = self._get_from_environ(name, ma.missing)
         self._fields[parsed_key] = field
+        source_key = proxied_key or parsed_key
         if raw_value is ma.missing and field.missing is ma.missing:
             message = "Environment variable not set."
             if self.eager:
-                raise EnvValidationError(
-                    'Environment variable "{}" not set'.format(proxied_key or parsed_key), [message]
-                )
+                raise EnvValidationError('Environment variable "{}" not set'.format(source_key), [message])
             else:
                 self._errors[parsed_key].append(message)
         if raw_value or raw_value == "":
@@ -78,7 +85,7 @@ def _field2method(
         except ma.ValidationError as error:
             if self.eager:
                 raise EnvValidationError(
-                    'Environment variable "{}" invalid: {}'.format(name, error.args[0]), error.messages
+                    'Environment variable "{}" invalid: {}'.format(source_key, error.args[0]), error.messages
                 ) from error
             self._errors[parsed_key].extend(error.messages)
         else:
@@ -108,18 +115,19 @@ def _func2method(func: typing.Callable, method_name: str) -> ParserMethod:
 
 
 # From webargs
-def _dict2schema(dct: typing.Dict[str, ma.fields.Field]) -> typing.Type[ma.Schema]:
+def _dict2schema(dct, schema_class=ma.Schema):
     """Generate a `marshmallow.Schema` class given a dictionary of
     `Fields <marshmallow.fields.Field>`.
     """
+    if hasattr(schema_class, "from_dict"):  # marshmallow 3
+        return schema_class.from_dict(dct)
     attrs = dct.copy()
-    if MARSHMALLOW_VERSION_INFO[0] < 3:
 
-        class Meta:
-            strict = True
+    class Meta:
+        strict = True
 
-        attrs["Meta"] = Meta
-    return type("", (ma.Schema,), attrs)
+    attrs["Meta"] = Meta
+    return type("", (schema_class,), attrs)
 
 
 def _make_list_field(*, subcast: Subcast, **kwargs) -> ma.fields.List:
@@ -132,8 +140,8 @@ def _preprocess_list(value: typing.Union[str, typing.Iterable], **kwargs) -> typ
 
 
 def _preprocess_dict(
-    value: typing.Union[str, typing.Mapping[str, T]], subcast: Subcast, **kwargs
-) -> typing.Mapping[str, T]:
+    value: typing.Union[str, typing.Mapping[str, _T]], subcast: Subcast, **kwargs
+) -> typing.Mapping[str, _T]:
     if isinstance(value, Mapping):
         return value
 
@@ -202,47 +210,45 @@ class Env:
 
     __call__ = _field2method(ma.fields.Field, "__call__")  # type: ParserMethod
 
-    default_parser_map = dict(
-        bool=_field2method(ma.fields.Bool, "bool"),
-        str=_field2method(ma.fields.Str, "str"),
-        int=_field2method(ma.fields.Int, "int"),
-        float=_field2method(ma.fields.Float, "float"),
-        decimal=_field2method(ma.fields.Decimal, "decimal"),
-        list=_field2method(_make_list_field, "list", preprocess=_preprocess_list),
-        dict=_field2method(ma.fields.Dict, "dict", preprocess=_preprocess_dict),
-        json=_field2method(ma.fields.Field, "json", preprocess=_preprocess_json),
-        datetime=_field2method(ma.fields.DateTime, "datetime"),
-        date=_field2method(ma.fields.Date, "date"),
-        path=_field2method(PathField, "path"),
-        log_level=_field2method(LogLevelField, "log_level"),
-        timedelta=_field2method(ma.fields.TimeDelta, "timedelta"),
-        uuid=_field2method(ma.fields.UUID, "uuid"),
-        url=_field2method(URLField, "url"),
-        dj_db_url=_func2method(_dj_db_url_parser, "dj_db_url"),
-        dj_email_url=_func2method(_dj_email_url_parser, "dj_email_url"),
-    )  # type: typing.Dict[str, ParserMethod]
+    int = _field2method(ma.fields.Int, "int")
+    bool = _field2method(ma.fields.Bool, "bool")
+    str = _field2method(ma.fields.Str, "str")
+    float = _field2method(ma.fields.Float, "float")
+    decimal = _field2method(ma.fields.Decimal, "decimal")
+    list = _field2method(_make_list_field, "list", preprocess=_preprocess_list)
+    dict = _field2method(ma.fields.Dict, "dict", preprocess=_preprocess_dict)
+    json = _field2method(ma.fields.Field, "json", preprocess=_preprocess_json)
+    datetime = _field2method(ma.fields.DateTime, "datetime")
+    date = _field2method(ma.fields.Date, "date")
+    path = _field2method(PathField, "path")
+    log_level = _field2method(LogLevelField, "log_level")
+    timedelta = _field2method(ma.fields.TimeDelta, "timedelta")
+    uuid = _field2method(ma.fields.UUID, "uuid")
+    url = _field2method(URLField, "url")
+    dj_db_url = _func2method(_dj_db_url_parser, "dj_db_url")
+    dj_email_url = _func2method(_dj_email_url_parser, "dj_email_url")
 
-    def __init__(self, *, eager: bool = True):
+    def __init__(self, *, eager: _BoolType = True):
         self.eager = eager
         self._sealed = False  # type: bool
-        self._fields = {}  # type: typing.Dict[str, ma.fields.Field]
-        self._values = {}  # type: typing.Dict[str, typing.Any]
+        self._fields = {}  # type: typing.Dict[_StrType, ma.fields.Field]
+        self._values = {}  # type: typing.Dict[_StrType, typing.Any]
         self._errors = collections.defaultdict(list)  # type: ErrorMapping
-        self._prefix = None  # type: typing.Optional[str]
-        self.__parser_map__ = self.default_parser_map.copy()
+        self._prefix = None  # type: typing.Optional[_StrType]
+        self.__custom_parsers__ = {}  # type: typing.Dict[_StrType, ParserMethod]
 
-    def __repr__(self) -> str:
+    def __repr__(self) -> _StrType:
         return "<{} {}>".format(self.__class__.__name__, self._values)
 
     __str__ = __repr__
 
     @staticmethod
     def read_env(
-        path: str = None,
-        recurse: bool = True,
-        stream: str = None,
-        verbose: bool = False,
-        override: bool = False,
+        path: _StrType = None,
+        recurse: _BoolType = True,
+        stream: _StrType = None,
+        verbose: _BoolType = False,
+        override: _BoolType = False,
     ) -> DotEnv:
         """Read a .env file into os.environ.
 
@@ -258,12 +264,15 @@ class Env:
                 raise RuntimeError("Could not get current call frame.")
             frame = current_frame.f_back
             caller_dir = os.path.dirname(frame.f_code.co_filename)
+            # Will be a directory
             start = os.path.join(os.path.abspath(caller_dir))
         else:
+            # Could be directory or a file
             start = path
         if recurse:
+            env_name = os.path.basename(start) if os.path.isfile(start) else ".env"
             for dirname in _walk_to_root(start):
-                check_path = os.path.join(dirname, ".env")
+                check_path = os.path.join(dirname, env_name)
                 if os.path.exists(check_path):
                     return load_dotenv(check_path, stream=stream, verbose=verbose, override=override)
         else:
@@ -272,7 +281,7 @@ class Env:
             return load_dotenv(start, stream=stream, verbose=verbose, override=override)
 
     @contextlib.contextmanager
-    def prefixed(self, prefix: str) -> typing.Iterator["Env"]:
+    def prefixed(self, prefix: _StrType) -> typing.Iterator["Env"]:
         """Context manager for parsing envvars with a common prefix."""
         try:
             old_prefix = self._prefix
@@ -299,20 +308,24 @@ class Env:
                 "Environment variables invalid: {}".format(error_messages), error_messages
             )
 
-    def __getattr__(self, name: str, **kwargs):
+    def __getattr__(self, name: _StrType, **kwargs):
         try:
-            return functools.partial(self.__parser_map__[name], self)
+            return functools.partial(self.__custom_parsers__[name], self)
         except KeyError as error:
             raise AttributeError("{} has no attribute {}".format(self, name)) from error
 
-    def add_parser(self, name: str, func: typing.Callable) -> None:
+    def add_parser(self, name: _StrType, func: typing.Callable) -> None:
         """Register a new parser method with the name ``name``. ``func`` must
         receive the input value for an environment variable.
         """
-        self.__parser_map__[name] = _func2method(func, method_name=name)
+        if hasattr(self, name):
+            raise ParserConflictError(
+                "Env already has a method with name '{}'. Use a different name.".format(name)
+            )
+        self.__custom_parsers__[name] = _func2method(func, method_name=name)
         return None
 
-    def parser_for(self, name: str) -> typing.Callable[[typing.Callable], typing.Callable]:
+    def parser_for(self, name: _StrType) -> typing.Callable[[typing.Callable], typing.Callable]:
         """Decorator that registers a new parser method with the name ``name``.
         The decorated function must receive the input value for an environment variable.
         """
@@ -323,11 +336,11 @@ class Env:
 
         return decorator
 
-    def add_parser_from_field(self, name: str, field_cls: typing.Type[ma.fields.Field]):
+    def add_parser_from_field(self, name: _StrType, field_cls: typing.Type[ma.fields.Field]):
         """Register a new parser method with name ``name``, given a marshmallow ``Field``."""
-        self.__parser_map__[name] = _field2method(field_cls, method_name=name)
+        self.__custom_parsers__[name] = _field2method(field_cls, method_name=name)
 
-    def dump(self) -> typing.Mapping[str, typing.Any]:
+    def dump(self) -> typing.Mapping[_StrType, typing.Any]:
         """Dump parsed environment variables to a dictionary of simple data types (numbers
         and strings).
         """
@@ -336,8 +349,8 @@ class Env:
         return dump_result.data if MARSHMALLOW_VERSION_INFO[0] < 3 else dump_result
 
     def _get_from_environ(
-        self, key: str, default: typing.Any, *, proxied: bool = False
-    ) -> typing.Tuple[str, typing.Any, typing.Optional[str]]:
+        self, key: _StrType, default: typing.Any, *, proxied: _BoolType = False
+    ) -> typing.Tuple[_StrType, typing.Any, typing.Optional[_StrType]]:
         """Access a value from os.environ. Handles proxied variables, e.g. SMTP_LOGIN={{MAILGUN_LOGIN}}.
 
         Returns a tuple (envvar_key, envvar_value, proxied_key). The ``envvar_key`` will be different from
@@ -355,5 +368,5 @@ class Env:
                 return (key, self._get_from_environ(proxied_key, default, proxied=True)[1], proxied_key)
         return env_key, value, None
 
-    def _get_key(self, key: str, *, omit_prefix: bool = False) -> str:
+    def _get_key(self, key: _StrType, *, omit_prefix: _BoolType = False) -> _StrType:
         return self._prefix + key if self._prefix and not omit_prefix else key