+++ CHANGELOG.md
@@ -1,8 +1,49 @@
 # Changelog
 
-## 5.2.0 (unreleased)
+## 6.0.0 (2019-09-22)
+
+Features:
+
+- Default parser methods are now defined as bound methods.
+  This enables static analysis features, e.g. autocomplete ([#103](https://github.com/sloria/environs/issues/103)).
+  Thanks [rugleb](https://github.com/rugleb) for the suggestion.
+  _Backwards-incompatible_: As a result of this change, adding a parser name that is the same as an existing method
+  will result in an error being raised.
+
+```python
+import environs
+
+env = environs.Env()
+
+# Below conflicts with built-in `url` method.
+# In <6.0.0, this would override the built-in method.
+# In >=6.0.0, this raises an error:
+#   environs.ParserConflictError: Env already has a method with name 'url'. Use a different name.
+@env.parser_for("url")
+def https_url(value):
+    return "https://" + value
+```
+
+Bug fixes:
+
+- Fix error message for prefixed variables ([#102](https://github.com/sloria/environs/issues/102)).
+  Thanks [AGeekInside](https://github.com/AGeekInside) for reporting.
 
 Other changes:
+
+- _Backwards-incompatible_: Rename `Env.__parser_map__` to `Env.__custom_parsers__`.
+
+## 5.2.1 (2019-08-08)
+
+Bug fixes:
+
+- Fix behavior when recursively searching for a specified file
+  ([#96](https://github.com/sloria/environs/issues/96)).
+  Thanks [ribeaud](https://github.com/ribeaud) for the catch and patch.
+
+## 5.2.0 (2019-07-19)
+
+Changes:
 
 - Improve typings.
 
@@ -153,3 +194,7 @@ Bug fixes:
 ## 0.1.0 (2016-04-25)
 
 - First PyPI release.
+
+```
+
+```